Fixing errors
Fatfingered an "o" at line 43. Was wondering why the script was failing...

+++ firstboot_trigger.sh
@@ -16,6 +16,8 @@ function _ask {
 }
 
 function main {
+    [[ ! -t 0 ]] && return 0;
+
     if [[ -f /var/lib/artix-firstboot-done ]]; then
         [[ "${EUID}" -eq 0 ]] && rm -f /etc/profile.d/firstboot.sh 2>/dev/null;
         return 0;
@@ -38,10 +40,11 @@ function main {
     printf "=======================================\n\n";
 
     if _ask "Run setup now?" "y"; then
-        printf "[*] Launching firstboot script...\n";o
+        printf "[*] Launching firstboot script...\n";
         if [[ "${EUID}" -eq 0 ]]; then
             exec /usr/local/bin/firstboot.sh;
         else
+            # Ovo ce traziti lozinku tvog usera ili roota
             exec sudo /usr/local/bin/firstboot.sh;
         fi
     else
@@ -50,6 +53,5 @@ function main {
         printf "    or run the wizard later from /usr/local/bin/firstboot.sh\n";
     fi
 }
-
 
 main;
